requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
pillow==9.0.0
numpy==1.22.0
requests==2.27.1
python-dotenv==0.19.2
google-generativeai==0.3.1
